Introduce a view class
Move all the console interaction into a view class making the game logic
isolated from UI.
