+++ 51 Hurkle/csharp/src/hurkle/GuessViewModel.cs
@@ -1,0 +1,7 @@
+namespace hurkle
+{
+    internal class GuessViewModel
+    {
+        public int CurrentGuessNumber {get;init;}
+    }
+}
