+++ 51 Hurkle/csharp/src/hurkle/IHurkleView.cs
@@ -1,0 +1,10 @@
+namespace hurkle
+{
+    internal interface IHurkleView
+    {
+        GamePoint GetGuess(GuessViewModel guessViewModel);
+        void ShowVictory(VictoryViewModel victoryViewModel);
+        void ShowDirection(FailedGuessViewModel failedGuessViewModel);
+        void ShowLoss(LossViewModel lossViewModel);
+    }
+}
